Create app.js which features a function to initialize the app using an API (mocked or real).
